+++ smdebug/mxnet/hook.py
@@ -5,7 +5,6 @@ import mxnet as mx
 from smdebug.core.collection import DEFAULT_MXNET_COLLECTIONS, CollectionKeys
 from smdebug.core.hook import CallbackHook
 from smdebug.core.json_config import DEFAULT_WORKER_NAME
-from smdebug.core.utils import is_first_process
 from smdebug.mxnet.collection import CollectionManager
 from smdebug.mxnet.graph import _net2pb
 from smdebug.mxnet.singleton_utils import set_hook
@@ -53,6 +52,7 @@ class Hook(CallbackHook):
         )
         self.last_block = None
         self.first_process = None
+        self._is_suppported_dist_strategy()
 
         self.model = None
         self.exported_model = False
@@ -61,20 +61,7 @@ class Hook(CallbackHook):
         self.worker = self._get_worker_name()
         set_hook(self)
 
-    def _is_suppported_dist_strategy(self):
-        num_workers = self._get_num_workers()
-        if num_workers > 1:
-            return True
-        else:
-            if self.first_process is not None:
-                self.first_process = is_first_process(self.out_dir)
-                self.save_all_workers = False
-                self.worker = self._get_worker_name()
-            return self.first_process
-
     def _get_worker_name(self):
-        if self.worker is not None:
-            return self.worker
         try:
             import horovod.mxnet as hvd
 
@@ -82,6 +69,8 @@ class Hook(CallbackHook):
                 return f"worker_{hvd.rank()}"
         except (ModuleNotFoundError, ValueError, ImportError):
             pass
+        if self.first_process is False:
+            return "non_chief_worker"
 
         return DEFAULT_WORKER_NAME
