+++ smdebug/pytorch/hook.py
@@ -8,7 +8,6 @@ import torch.distributed as dist
 from smdebug.core.collection import DEFAULT_PYTORCH_COLLECTIONS, CollectionKeys
 from smdebug.core.hook import CallbackHook
 from smdebug.core.json_config import DEFAULT_WORKER_NAME
-from smdebug.core.utils import is_first_process
 from smdebug.pytorch.collection import CollectionManager
 from smdebug.pytorch.singleton_utils import set_hook
 from smdebug.pytorch.utils import get_reduction_of_data, make_numpy_array
@@ -51,22 +50,11 @@ class Hook(CallbackHook):
         self.module_set = set()
 
         self.first_process = None
+        self._is_suppported_dist_strategy()
         self.has_registered_module = False
         self.has_registered_loss_module = False
         self.worker = self._get_worker_name()
-        self.first_process = self._is_suppported_dist_strategy()
         set_hook(self)
-
-    def _is_suppported_dist_strategy(self):
-        num_workers = self._get_num_workers()
-        if num_workers > 1:
-            return True
-        else:
-            if self.first_process is not None:
-                self.first_process = is_first_process(self.out_dir)
-                self.save_all_workers = False
-                self.worker = self._get_worker_name()
-            return self.first_process
 
     def _get_num_workers(self):
         """Check horovod and torch.distributed."""
@@ -88,8 +76,6 @@ class Hook(CallbackHook):
         return 1
 
     def _get_worker_name(self):
-        if self.worker is not None:
-            return self.worker
         """Check horovod and torch.distributed."""
         # Try torch.distributed
         # torch.distributed is empty on Mac on Torch <= 1.2
@@ -104,9 +90,7 @@ class Hook(CallbackHook):
                     return f"worker_{hvd.rank()}"
             except (ModuleNotFoundError, ValueError, ImportError):
                 pass
-
-        if not is_first_process(self.out_dir):
-            self.save_all_workers = False
+        if self.first_process is False:
             return "non_chief_worker"
 
         # Return default
